backend/requirements.txt
# =============================================================================
# AgriLite-Hybrid Backend
# requirements.txt - Python Dependencies
# 
# Install with: pip install -r requirements.txt
# =============================================================================

# Core Framework
Flask==3.0.0
python-dotenv==1.0.0

# Database & ORM
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication & Security
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
cryptography==41.0.7

# Session Management
Flask-Session==0.5.0
cachelib==0.13.0

# ML / Deep Learning
tensorflow==2.16.1
numpy==1.26.4
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# Image Processing
scikit-image==0.22.0

# Utilities
Werkzeug==3.0.1

# Testing
pytest==8.4.2
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development (optional)
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Production (optional - for Redis caching)
# redis==5.0.1

# Supabase (optional - for cloud database)
# supabase==2.3.4
